requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via flask
elasticsearch==5.4.0
flask-admin==1.5.0
flask-login==0.4.0
flask-wtf==0.14.2
flask==0.12.2
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
markupsafe==1.0           # via jinja2
neo4j-driver==1.5.0
sqlineage==0.2.1
urllib3==1.22             # via elasticsearch
urlparse3==1.1
werkzeug==0.12.2          # via flask
wtforms==2.1              # via flask-admin, flask-wtf
